# api/openapi.yml
openapi: 3.0.3
info:
  title: ym music service
  version: "0.1.0"
paths:
...
  /user/{userId}:
    get:
      summary: get user by id
      operationId: getUserById
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int32
          required: true
          description: user id, represented as int
      responses:
        '200':
          description: User data fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserReturn'
        '401':
          description: Unauthorized access, if something goes wrong with tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User with this id not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: update user
      operationId: changeUser
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int32
          required: true
          description: user id, represented as int
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserReturn'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/{userId}/avatar:
    get:
      summary: get uploaded avatar
      operationId: getUserAvatar
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int32
          required: true
      responses:
        '200':
          description: avatar was found
          content:
            image/webp:
              schema:
                type: string
                format: binary
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: avatar wasn't found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: upload new user's avatar to server
      operationId: uploadUserAvatar
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int32
          required: true
      requestBody:
        required: true
        content:
          image/webp:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: avatar was successfully uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: wrong format of an image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: delete current avatar
      operationId: deleteUserAvatar
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int32
          required: true
      responses:
        '200':
          description: avatar was successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/{userId}/change_password:
    patch:
      summary: change user's password
      operationId: changePassword
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int32
          required: true
          description: user id, represented as int
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserChangePassword'
      responses:
        '200':
          description: User's password was successfully updated
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid old password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /track:
    post:
      summary: uploads new track, make all transcoding stuff, stores it
      operationId: uploadTrack
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TrackUploadRequest'
            encoding:
              track:
                contentType: audio/*
      responses:
        '200':
          description: Track uploaded and processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackUploadSuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type, uploading file is not an audio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: get all uploaded tracks meta
      operationId: getTracks
      responses:
        '200':
          description: tracks was successfully fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrackMetadata'
        '404':
          description: unkown userId
  /track/{trackId}:
    delete:
      summary: delete track
      operationId: deleteTrack
      parameters:
        - in: path
          name: trackId
          schema:
            type: integer
            format: int32
          required: true
          description: track id, represented as int
      responses:
        '200':
          description: Track was successfully deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such Track
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: getting track metadata
      operationId: getTrackMeta
      parameters:
        - in: path
          name: trackId
          schema:
            type: integer
            format: int32
          required: true
          description: track id, represented as int
      responses:
        '200':
          description: track meta was successfully fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackMetadata'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: no track found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /track/{id}/stream:
    head:
      summary: getting track metadata, needed for streaming
      operationId: streamTrackHead
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int32
          required: true
        - in: query
          name: quality
          schema:
            type: string
          required: false
      responses: 
        '200':
          description: "metadata was fetched"
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "no track found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: getting track stream
      operationId: streamTrack
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int32
          required: true
        - in: query
          name: quality
          schema:
            type: string
          required: false
      responses:
        '200':
          description: "start streaming"
          content:
            audio/*:
              schema:
                type: string
                format: binary
        '206':
          description: "partial"
          content:
            audio/*:
              schema:
                type: string
                format: binary
        '416':
          description: "invalid range"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "no track found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserReturn:
      type: object
      required:
        - username
        - email
        - id
      properties:
        id:
          type: integer
          format: int32
        username:
          type: string
        email:
          type: string
          nullable: true
    UserChangePassword:
      type: object
      required:
        - old_password
        - new_password
      properties:
        old_password:
          type: string
        new_password:
          type: string
    UserUpdate:
      type: object
      required:
        - new_username
        - new_email
      properties:
        new_username:
          type: string
        new_email:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string
    TrackUploadSuccessResponse:
      type: object
      required:
        - track_id
      properties:
        track_id:
          type: integer
          format: int32
    TrackUploadRequest:
      type: object
      required:
        - name
        - album_id
        - artist_id
        - track
      properties:
        name:
          type: string
        album_id:
          type: integer
          format: int32
        artist_id:
          type: integer
          format: int32
        track:
          type: string
          format: binary
    TrackMetadata:
      type: object
      required:
        - track_id
        - name
        - artist_id
      properties:
        track_id:
          type: integer
          format: int32
        name:
          type: string
        artist_id:
          type: integer
          format: int32
        cover_url:
          type: string
        track_fast_preset:
          type: string
        track_standard_preset:
          type: string
        track_high_preset:
          type: string
        track_lossless_preset:
          type: string
...
security:
  - bearerAuth: []